+++ web/src/layouts/mod.rs
@@ -1,4 +1,3 @@
 pub mod authenticated;
-pub mod email_validated;
 pub mod fullpage;
 pub mod nav_bar;